PDF Viewer Basic
pdf viewer
concepts Learnt:
<form> 
<worker>, <viewer> 
FileReader();
FileReader.readAsdataUrl();
FileReader.onloadend();
FileReader.result..

-----------------------
Explanation:
Form:
The <form> tag is used to create an HTML form for user input.
The <form> element can contain one or more of the following form elements:
<input>
<textarea>
<button>
<select>
<option>
<optgroup>
<fieldset>
<label>
<output>

Onsubmit: The onsubmit attribute fires when a form is submitted.

<Worker> Component:
This component is likely from @react-pdf-viewer. The workerUrl prop specifies the URL of the PDF worker script, which is responsible for parsing and rendering the PDF documents in a separate thread to avoid blocking the main thread.
https://unpkg.com/pdfjs-dist@2.16.105/build/pdf.worker.min.js is the URL where the worker script is hosted.

<Viewer> Component:
This is the main component responsible for displaying the PDF. The fileUrl prop specifies the URL of the PDF file to be viewed.
The plugins prop accepts an array of plugins that enhance the functionality of the viewer. In this case, defaultLayoutPluginInstance is passed to provide default layout features.

FileReader: readAsDataURL() method
The readAsDataURL() method of the FileReader interface is used to read the contents of the specified Blob or File. When the read operation is finished, the readyState property becomes DONE, and the loadend event is triggered. At that time, the result attribute contains the data as a data: URL representing the file's data as a base64 encoded string.

FileReader: loadend event
The loadend event of the FileReader interface is fired when a file read has completed, successfully or not.
This event is not cancelable and does not bubble.